requirements.txt
# requirements.txt

# 核心 Discord 库
discord.py

# discord.py 的一个重要依赖，通常会自动安装，但明确列出是好习惯
aiohttp

aiofiles

pytz
python-dotenv
pydantic>=2.0

# 系统监测
psutil
distro


rich